# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridWitness ingest",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_v1_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/samples": {
      "post": {
        "summary": "Post Samples",
        "operationId": "post_samples_v1_samples_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/node": {
      "patch": {
        "summary": "Update Node",
        "operationId": "update_node_v1_node_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Node V1 Node Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Node",
        "operationId": "delete_node_v1_node_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Node V1 Node Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/time-echo": {
      "post": {
        "summary": "Time Echo",
        "operationId": "time_echo_v1_time_echo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeEchoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeEchoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health V1 Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConsentUpdate": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "loc_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "postcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postcode"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "ConsentUpdate"
      },
      "DeleteRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "DeleteRequest"
      },
      "ElectricalRow": {
        "properties": {
          "ts_utc": {
            "type": "string",
            "title": "Ts Utc"
          },
          "ts_source": {
            "type": "string",
            "enum": [
              "device",
              "ha_receive",
              "gps"
            ],
            "title": "Ts Source",
            "default": "ha_receive"
          },
          "phase": {
            "type": "string",
            "enum": [
              "L1",
              "L2",
              "L3",
              "1p"
            ],
            "title": "Phase",
            "default": "1p"
          },
          "voltage_v": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voltage V"
          },
          "current_a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current A"
          },
          "power_w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power W"
          },
          "power_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Factor"
          },
          "frequency_hz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency Hz"
          },
          "phase_angle_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase Angle Deg"
          }
        },
        "type": "object",
        "required": [
          "ts_utc"
        ],
        "title": "ElectricalRow"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocTier": {
        "type": "string",
        "enum": [
          "anon",
          "region",
          "data_share"
        ],
        "title": "LocTier"
      },
      "RegisterRequest": {
        "properties": {
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          },
          "loc_tier": {
            "$ref": "#/components/schemas/LocTier",
            "default": "anon"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "postcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postcode"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type",
            "default": "unknown"
          },
          "firmware": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware"
          },
          "cadence_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cadence Ms"
          },
          "producer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          }
        },
        "type": "object",
        "title": "RegisterRequest"
      },
      "RegisterResponse": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "loc_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loc Ref"
          },
          "cell_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Id"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "token",
          "loc_ref",
          "cell_id"
        ],
        "title": "RegisterResponse"
      },
      "RejectedRow": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "kind": {
            "type": "string",
            "enum": [
              "electrical",
              "weather"
            ],
            "title": "Kind"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "index",
          "kind",
          "reason"
        ],
        "title": "RejectedRow"
      },
      "SamplesRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "client_send_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Send Ts"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "electrical": {
            "items": {
              "$ref": "#/components/schemas/ElectricalRow"
            },
            "type": "array",
            "title": "Electrical"
          },
          "weather": {
            "items": {
              "$ref": "#/components/schemas/WeatherRow"
            },
            "type": "array",
            "title": "Weather"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "SamplesRequest"
      },
      "SamplesResponse": {
        "properties": {
          "server_receive_ts": {
            "type": "string",
            "title": "Server Receive Ts"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedRow"
            },
            "type": "array",
            "title": "Rejected"
          }
        },
        "type": "object",
        "required": [
          "server_receive_ts",
          "accepted",
          "duplicates"
        ],
        "title": "SamplesResponse"
      },
      "TimeEchoRequest": {
        "properties": {
          "client_send": {
            "type": "string",
            "title": "Client Send"
          }
        },
        "type": "object",
        "required": [
          "client_send"
        ],
        "title": "TimeEchoRequest"
      },
      "TimeEchoResponse": {
        "properties": {
          "client_send": {
            "type": "string",
            "title": "Client Send"
          },
          "server_receive": {
            "type": "string",
            "title": "Server Receive"
          }
        },
        "type": "object",
        "required": [
          "client_send",
          "server_receive"
        ],
        "title": "TimeEchoResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeatherRow": {
        "properties": {
          "time": {
            "type": "string",
            "title": "Time"
          },
          "ts_source": {
            "type": "string",
            "enum": [
              "device",
              "ha_receive"
            ],
            "title": "Ts Source",
            "default": "ha_receive"
          },
          "temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp"
          },
          "rhum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rhum"
          },
          "wspd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wspd"
          },
          "wdir": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wdir"
          },
          "pres": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pres"
          },
          "prcp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prcp"
          },
          "solar_radiation_w_m2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solar Radiation W M2"
          },
          "uv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uv"
          }
        },
        "type": "object",
        "required": [
          "time"
        ],
        "title": "WeatherRow"
      }
    }
  }
}